rename projectToPlane in dir2 dir3 dist2 dist3
