+++ lib/box2/projectToPlane.js
@@ -2,7 +2,8 @@ const transitTo = require('./transitTo')
 const homothety = require('./homothety')
 
 module.exports = (box, target, camera) => {
-  // @affineplane.box2.projectTo(box, target[, camera])
+  // @affineplane.box2.projectToPlane(box, target[, camera])
+  // @affineplane.box2.projectTo
   //
   // Project a box onto a target plane in 3d.
   // If a camera is given, project perspectively.
