+++ lib/box3/projectToPlane.js
@@ -2,7 +2,8 @@ const transitTo = require('../box2/transitTo')
 const homothety = require('../box2/homothety')
 
 module.exports = (box, target, camera) => {
-  // @affineplane.box3.projectTo(box, target[, camera])
+  // @affineplane.box3.projectToPlane(box, target[, camera])
+  // @affineplane.box3.projectTo
   //
   // Project a 3D box onto a target plane.
   // If a camera is given, project perspectively.
