+++ lib/dir2/index.js
@@ -11,8 +11,8 @@ exports.create = require('./fromPolar')
 exports.fromAngle = exports.create
 exports.fromPolar = exports.create
 exports.fromVector = require('./fromVector')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.toAngle = require('./toPolar')
 exports.toPolar = exports.toAngle
 exports.toVector = require('./toVector')
