+++ lib/dir2/projectToPlane.js
@@ -1,4 +1,5 @@
-// @affineplane.dir2.projectTo(dir, plane)
+// @affineplane.dir2.projectToPlane(dir, plane)
+// @affineplane.dir2.projectTo
 //
 // Project a 2D direction onto a 2D plane. Perspective does not
 // affect the direction.
