+++ lib/dir3/index.js
@@ -9,8 +9,8 @@ exports.almostEqual = require('./almostEqual')
 exports.create = require('./fromSpherical')
 exports.fromSpherical = exports.create
 exports.fromVector = require('./fromVector')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.transitFrom = require('./transitFrom')
 exports.transitTo = require('./transitTo')
 exports.toSpherical = require('./toSpherical')
