+++ lib/dir3/projectToPlane.js
@@ -2,7 +2,8 @@ const transitTo = require('./transitTo')
 const dir2FromVector = require('../dir2/fromVector')
 
 module.exports = (dir, plane) => {
-  // @affineplane.dir3.projectTo(dir, plane)
+  // @affineplane.dir3.projectToPlane(dir, plane)
+  // @affineplane.dir3.projectTo
   //
   // Project a 3D direction onto a 2D plane orthogonally.
   // We cannot project 3D directions perspectively because
