+++ lib/dist2/projectToPlane.js
@@ -1,7 +1,8 @@
 const transitTo = require('./transitTo')
 
 module.exports = (dist, target, camera) => {
-  // @affineplane.dist2.projectTo(dist, target[, camera])
+  // @affineplane.dist2.projectToPlane(dist, target[, camera])
+  // @affineplane.dist2.projectTo
   //
   // Project a distance onto a target plane in 3D.
   // The distance is assumed to be measured on the reference plane z=0.
