+++ lib/dist3/index.js
@@ -11,7 +11,7 @@
 exports.almostEqual = require('./almostEqual')
 exports.create = require('./create')
 exports.equal = require('./equal')
-exports.projectTo = require('./projectTo')
-exports.projectToPlane = exports.projectTo
+exports.projectToPlane = require('./projectToPlane')
+exports.projectTo = exports.projectToPlane
 exports.transitFrom = require('./transitFrom')
 exports.transitTo = require('./transitTo')
