+++ lib/dist3/projectToPlane.js
@@ -1,4 +1,5 @@
-// @affineplane.dist3.projectTo(dist, target[, camera])
+// @affineplane.dist3.projectToPlane(dist, target[, camera])
+// @affineplane.dist3.projectTo
 //
 // Project a distance onto a target plane in 3D basis.
 // The distance is assumed to be measured
@@ -21,4 +22,4 @@
 //
 
 // Reuse because same assumptions
-module.exports = require('../dist2/projectTo')
+module.exports = require('../dist2/projectToPlane')
